% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/bbd_prob.R
\name{bbd_prob}
\alias{bbd_prob}
\title{Transition probabilities of a birth/birth-death process}
\usage{
bbd_prob(t, a0, b0, lambda1, lambda2, mu2, gamma, A, B, nblocks = 256,
  tol = 1e-12, computeMode = 0, nThreads = 4, maxdepth = 400,
  vec_output = NULL)
}
\arguments{
\item{t}{time}

\item{a0}{total number of type 1 particles at \code{t = 0}}

\item{b0}{total number of type 2 particles at \code{t = 0}}

\item{lambda1}{birth rate of type 1 particles (a two variables function)}

\item{lambda2}{birth rate of type 2 particles (a two variables function)}

\item{mu2}{death rate function of type 2 particles (a two variables function)}

\item{gamma}{transition rate from type 2 particles to type 1 particles (a two variables function)}

\item{A}{upper bound for the total number of type 1 particles}

\item{B}{upper bound for the total number of type 2 particles}

\item{nblocks}{number of blocks}

\item{tol}{tolerance}

\item{computeMode}{computation mode}

\item{nThreads}{number of threads}

\item{maxdepth}{maximum number of iterations for Lentz algorithm}
}
\description{
Computes the transition pobabilities of a birth/birth-death process
using the continued fraction representation of its Laplace transform
}

