\name{MultiBD-package}
\alias{MultiBD-package}
\alias{MultiBD}
\docType{package}
\title{
Multivariate birth-death processes
}
\description{
computes the transition probabilities of several multivariate birth-death processes
}
\details{
\tabular{ll}{
Package: \tab MultiBD\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-20\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Lam Si Tung Ho, Forrest W. Crawford, Marc A. Suchard

Maintainer: Lam Si Tung Ho <lamho86@gmail.com>
}
\references{
Ho, L. S. T. et al. 2015. "Birth(death)/birth-death processes and their computable transition probabilities with statistical applications". In review.
}

