% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/MultiBD.R
\docType{package}
\name{MultiBD}
\alias{MultiBD}
\alias{MultiBD-package}
\title{Multivariate birth-death processes}
\description{
The MultiBD package computes the transition probabilities
of several multivariate birth-death processes.
}
\references{
Ho LST et al. 2015. "Birth(death)/birth-death processes and their computable transition probabilities with statistical applications". In review.
}

