% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBD.R
\docType{package}
\name{MultiBD}
\alias{MultiBD}
\alias{MultiBD-package}
\title{MultiBD: Cyclic coordinate descent for logistic, Poisson and survival analysis}
\description{
The MultiBD package incorporates cyclic coordinate descent and
majorization-minimization approaches to fit a variety of regression models
found in large-scale observational healthcare data.  Implementations focus
on computational optimization and fine-scale parallelization to yield
efficient inference in massive datasets.
}

